requirement.txt
customtkinter
python-docx
PyPDF2
openpyxl
pandas
GPT4All
pytesseract
pytesseract pillow